macos: remove the old split implementation
